消息缓存策略两个默认实现
